@@ -4,7 +4,7 @@ import android.content.Context
 import android.content.Intent
 import android.graphics.Typeface
 import android.os.Parcelable
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -25,7 +25,7 @@ import java.text.SimpleDateFormat
 class FeedAdapter(context: Context, list: ArrayList<HotBean.ItemListBean.DataBean>) : RecyclerView.Adapter<FeedAdapter.FeedViewHolder>() {
     var context: Context? = null;
     var list: ArrayList<HotBean.ItemListBean.DataBean>? = null
-    var inflater: LayoutInflater? = null
+    var inflater: LayoutInflater
 
     init {
         this.context = context
@@ -33,7 +33,7 @@ class FeedAdapter(context: Context, list: ArrayList<HotBean.ItemListBean.DataBea
         this.inflater = LayoutInflater.from(context)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): FeedViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FeedViewHolder {
         return FeedViewHolder(inflater?.inflate(R.layout.item_feed_result, parent, false), context!!)
     }
 
@@ -41,7 +41,7 @@ class FeedAdapter(context: Context, list: ArrayList<HotBean.ItemListBean.DataBea
         return list?.size ?: 0
     }
 
-    override fun onBindViewHolder(holder: FeedViewHolder?, position: Int) {
+    override fun onBindViewHolder(holder: FeedViewHolder, position: Int) {
         var photoUrl : String? = list?.get(position)?.cover?.feed
         photoUrl?.let { ImageLoadUtils.display(context!!,holder?.iv_photo, it) }
         var title : String? = list?.get(position)?.title
@@ -95,7 +95,7 @@ class FeedAdapter(context: Context, list: ArrayList<HotBean.ItemListBean.DataBea
     }
 
 
-    class FeedViewHolder(itemView: View?, context: Context) : RecyclerView.ViewHolder(itemView) {
+    class FeedViewHolder(itemView: View, context: Context) : RecyclerView.ViewHolder(itemView) {
         var iv_photo: ImageView = itemView?.findViewById(R.id.iv_photo) as ImageView
         var tv_title: TextView = itemView?.findViewById(R.id.tv_title) as TextView
         var tv_time: TextView = itemView?.findViewById(R.id.tv_detail) as TextView
