@@ -1,15 +1,13 @@
 package com.tt.lvruheng.eyepetizer.adapter
 
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
-import android.view.View
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentPagerAdapter
 
 /**
  * Created by lvruheng on 2017/7/6.
  */
-class HotAdatpter(fm: FragmentManager,list: ArrayList<Fragment>,titles : MutableList<String>) : FragmentPagerAdapter(fm) {
-    var mFm : FragmentManager = fm!!
+class HotAdatpter(fm: FragmentManager?, list: ArrayList<Fragment>, titles: MutableList<String>) : FragmentPagerAdapter(fm!!, BEHAVIOR_SET_USER_VISIBLE_HINT) {
     var mList : ArrayList<Fragment> = list
     var mTitles : MutableList<String> = titles
     override fun getItem(position: Int): Fragment {
