@@ -2,7 +2,7 @@ package com.tt.lvruheng.eyepetizer.adapter
 
 import android.content.Context
 import android.content.Intent
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -15,7 +15,7 @@ import com.tt.lvruheng.eyepetizer.ui.ResultActivity
 /**
  * Created by lvruheng on 2017/7/9.
  */
-class SearchAdapter(context: Context, list: ArrayList<String>) : RecyclerView.Adapter<SearchAdapter.SearchViewHolder>() {
+class SearchAdapter(context: Context?, list: ArrayList<String>) : RecyclerView.Adapter<SearchAdapter.SearchViewHolder>() {
     var context: Context? = null;
     var list: ArrayList<String>? = null
     var inflater: LayoutInflater? = null
@@ -27,11 +27,11 @@ class SearchAdapter(context: Context, list: ArrayList<String>) : RecyclerView.Ad
         this.inflater = LayoutInflater.from(context)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): SearchViewHolder {
-        return SearchViewHolder(inflater?.inflate(R.layout.item_search, parent, false), context!!)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SearchViewHolder {
+        return SearchViewHolder(inflater!!.inflate(R.layout.item_search, parent, false), context!!)
     }
 
-    override fun onBindViewHolder(holder: SearchViewHolder?, position: Int) {
+    override fun onBindViewHolder(holder: SearchViewHolder, position: Int) {
         holder?.tv_title?.text = list!![position]
         val params = holder?.tv_title?.layoutParams
         if (params is FlexboxLayoutManager.LayoutParams) {
@@ -51,7 +51,7 @@ class SearchAdapter(context: Context, list: ArrayList<String>) : RecyclerView.Ad
     }
 
 
-    class SearchViewHolder(itemView: View?, context: Context) : RecyclerView.ViewHolder(itemView) {
+    class SearchViewHolder(itemView: View, context: Context) : RecyclerView.ViewHolder(itemView) {
         var tv_title: TextView = itemView?.findViewById(R.id.tv_title) as TextView
         
     }
