@@ -4,7 +4,7 @@ import android.content.Context
 import android.content.Intent
 import android.graphics.Typeface
 import android.os.Parcelable
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -22,10 +22,10 @@ import java.text.SimpleDateFormat
 /**
  * Created by lvruheng on 2017/7/7.
  */
-class WatchAdapter(context: Context, list: ArrayList<VideoBean>) : RecyclerView.Adapter<WatchAdapter.WatchViewHolder>() {
+class WatchAdapter(context: Context?, list: ArrayList<VideoBean>) : RecyclerView.Adapter<WatchAdapter.WatchViewHolder>() {
     var context: Context? = null;
     var list: ArrayList<VideoBean>? = null
-    var inflater: LayoutInflater? = null
+    var inflater: LayoutInflater
 
     init {
         this.context = context
@@ -33,15 +33,15 @@ class WatchAdapter(context: Context, list: ArrayList<VideoBean>) : RecyclerView.
         this.inflater = LayoutInflater.from(context)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): WatchViewHolder {
-        return WatchViewHolder(inflater?.inflate(R.layout.item_feed_result, parent, false), context!!)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WatchViewHolder {
+        return WatchViewHolder(inflater.inflate(R.layout.item_feed_result, parent, false), context!!)
     }
 
     override fun getItemCount(): Int {
         return list?.size ?: 0
     }
 
-    override fun onBindViewHolder(holder: WatchViewHolder?, position: Int) {
+    override fun onBindViewHolder(holder: WatchViewHolder, position: Int) {
         var photoUrl : String? = list?.get(position)?.feed
         photoUrl?.let { ImageLoadUtils.display(context!!,holder?.iv_photo, it) }
         var title : String? = list?.get(position)?.title
@@ -95,7 +95,7 @@ class WatchAdapter(context: Context, list: ArrayList<VideoBean>) : RecyclerView.
     }
 
 
-    class WatchViewHolder(itemView: View?, context: Context) : RecyclerView.ViewHolder(itemView) {
+    class WatchViewHolder(itemView: View, context: Context) : RecyclerView.ViewHolder(itemView) {
         var iv_photo: ImageView = itemView?.findViewById(R.id.iv_photo) as ImageView
         var tv_title: TextView = itemView?.findViewById(R.id.tv_title) as TextView
         var tv_time: TextView = itemView?.findViewById(R.id.tv_detail) as TextView
