@@ -1,10 +1,10 @@
 package com.tt.lvruheng.eyepetizer.ui.fragment
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.widget.SwipeRefreshLayout
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.fragment.app.Fragment
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.util.Log
 import android.util.Log.println
 import android.view.LayoutInflater
@@ -12,19 +12,19 @@ import android.view.View
 import android.view.ViewGroup
 import com.tt.lvruheng.eyepetizer.R
 import com.tt.lvruheng.eyepetizer.adapter.HomeAdatper
+import com.tt.lvruheng.eyepetizer.databinding.HomeFragmentBinding
 import com.tt.lvruheng.eyepetizer.mvp.contract.HomeContract
 import com.tt.lvruheng.eyepetizer.mvp.model.bean.HomeBean
 import com.tt.lvruheng.eyepetizer.mvp.model.bean.HomeBean.IssueListBean.ItemListBean
 import com.tt.lvruheng.eyepetizer.mvp.presenter.HomePresenter
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.home_fragment.*
+
 import java.util.*
 import java.util.regex.Pattern
 
 /**
  * Created by lvruheng on 2017/7/4.
  */
-class HomeFragment : BaseFragment(), HomeContract.View, SwipeRefreshLayout.OnRefreshListener {
+class HomeFragment : BaseFragment<HomeFragmentBinding>(), HomeContract.View, androidx.swiperefreshlayout.widget.SwipeRefreshLayout.OnRefreshListener {
     var mIsRefresh: Boolean = false
     var mPresenter: HomePresenter? = null
     var mList = ArrayList<ItemListBean>()
@@ -37,7 +37,7 @@ class HomeFragment : BaseFragment(), HomeContract.View, SwipeRefreshLayout.OnRef
         data = m.replaceAll("").subSequence(1, m.replaceAll("").length - 1).toString()
         if (mIsRefresh) {
             mIsRefresh = false
-            refreshLayout.isRefreshing = false
+            binding.refreshLayout.isRefreshing = false
             if (mList.size > 0) {
                 mList.clear()
             }
@@ -58,14 +58,14 @@ class HomeFragment : BaseFragment(), HomeContract.View, SwipeRefreshLayout.OnRef
     override fun initView() {
         mPresenter = HomePresenter(context, this)
         mPresenter?.start()
-        recyclerView.layoutManager = LinearLayoutManager(context)
+        binding.recyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(context)
         mAdapter = HomeAdatper(context, mList)
-        recyclerView.adapter = mAdapter
-        refreshLayout.setOnRefreshListener(this)
-        recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
-            override fun onScrollStateChanged(recyclerView: RecyclerView?, newState: Int) {
+        binding.recyclerView.adapter = mAdapter
+        binding.refreshLayout.setOnRefreshListener(this)
+        binding.recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+            override fun onScrollStateChanged(recyclerView: androidx.recyclerview.widget.RecyclerView, newState: Int) {
                 super.onScrollStateChanged(recyclerView, newState)
-                var layoutManager: LinearLayoutManager = recyclerView?.layoutManager as LinearLayoutManager
+                var layoutManager: androidx.recyclerview.widget.LinearLayoutManager = binding.recyclerView.layoutManager as androidx.recyclerview.widget.LinearLayoutManager
                 var lastPositon = layoutManager.findLastVisibleItemPosition()
                 if (newState == RecyclerView.SCROLL_STATE_IDLE && lastPositon == mList.size - 1) {
                     if (data != null) {
