@@ -13,14 +13,14 @@ import java.net.URLEncoder
  * Created by lvruheng on 2017/7/7.
  */
 class FindDetailModel {
-    fun loadData(context: Context, categoryName: String, strategy: String?): Observable<HotBean>? {
+    fun loadData(context: Context, categoryId: Int, strategy: String?): Observable<HotBean>? {
         val retrofitClient = RetrofitClient.getInstance(context, ApiService.BASE_URL)
         val apiService = retrofitClient.create(ApiService::class.java)
-        return apiService?.getFindDetailData(categoryName, strategy!!, "26868b32e808498db32fd51fb422d00175e179df", 83)
+        return apiService?.getFindDetailData(categoryId, strategy!!, "26868b32e808498db32fd51fb422d00175e179df", 83)
     }
-    fun loadMoreData(context: Context,start : Int, categoryName: String, strategy: String?): Observable<HotBean>? {
+    fun loadMoreData(context: Context,start : Int, categoryId: Int, strategy: String?): Observable<HotBean>? {
         val retrofitClient = RetrofitClient.getInstance(context, ApiService.BASE_URL)
         val apiService = retrofitClient.create(ApiService::class.java)
-        return apiService?.getFindDetailMoreData(start,10,categoryName, strategy!!)
+        return apiService?.getFindDetailMoreData(start,categoryId, strategy!!)
     }
 }
