@@ -28,7 +28,8 @@ interface ApiService{
     fun getHomeMoreData(@Query("date") date :String,@Query("num") num :String) : Observable<HomeBean>
 
     //获取发现频道信息
-    @GET("v2/categories?udid=26868b32e808498db32fd51fb422d00175e179df&vc=83")
+    //http://baobab.kaiyanapp.com/api/v4/categories/
+    @GET("v4/categories?udid=26868b32e808498db32fd51fb422d00175e179df&vc=83")
     fun getFindData() : Observable<MutableList<FindBean>>
 
 
@@ -38,14 +39,15 @@ interface ApiService{
                    @Query("udid") udid :String,@Query("vc") vc :Int) : Observable<HotBean>
 
     //获取发现频道详情信息
-    @GET("v3/videos")
-    fun getFindDetailData(@Query("categoryName") categoryName :String,@Query("strategy") strategy :String,
+    //http://baobab.kaiyanapp.com/api/v4/categories/detail/index?id=14
+    @GET("v4/categories/detail/index")
+    fun getFindDetailData(@Query("id") categoryId :Int,@Query("strategy") strategy :String,
                           @Query("udid") udid :String,@Query("vc") vc :Int) : Observable<HotBean>
 
     //获取发现详情加载更多消息
-    @GET("v3/videos")
-    fun getFindDetailMoreData(@Query("start") start :Int,@Query("num") num :Int,
-                              @Query("categoryName") categoryName :String,@Query("strategy") strategy :String) : Observable<HotBean>
+    @GET("v4/categories/detail/index")
+    fun getFindDetailMoreData(@Query("start") start :Int,
+                              @Query("id") categoryId :Int,@Query("strategy") strategy :String) : Observable<HotBean>
     //获取关键词搜索相关信息
     @GET("v1/search")
     fun getSearchData(@Query("num") num :Int,@Query("query") query :String,
