@@ -4,19 +4,19 @@ import android.app.DialogFragment
 import android.content.DialogInterface
 import android.content.Intent
 import android.graphics.Typeface
-import com.tt.lvruheng.eyepetizer.R
 import android.os.Bundle
-import android.view.*
-import com.tt.lvruheng.eyepetizer.utils.KeyBoardUtils
-import kotlinx.android.synthetic.main.search_fragment.*
-import android.widget.Toast
+import androidx.recyclerview.widget.DefaultItemAnimator
 import android.text.TextUtils
-import com.tt.lvruheng.eyepetizer.adapter.SearchAdapter
-import android.support.v7.widget.DefaultItemAnimator
-import com.google.android.flexbox.FlexWrap
+import android.view.*
+import android.widget.Toast
 import com.google.android.flexbox.FlexDirection
+import com.google.android.flexbox.FlexWrap
 import com.google.android.flexbox.FlexboxLayoutManager
+import com.tt.lvruheng.eyepetizer.R
+import com.tt.lvruheng.eyepetizer.adapter.SearchAdapter
+import com.tt.lvruheng.eyepetizer.databinding.SearchFragmentBinding
 import com.tt.lvruheng.eyepetizer.ui.ResultActivity
+import com.tt.lvruheng.eyepetizer.utils.KeyBoardUtils
 
 
 /**
@@ -30,9 +30,13 @@ class SearchFragment : DialogFragment(), CircularRevealAnim.AnimListener,
     var data : MutableList<String> = arrayListOf("脱口秀","城会玩","666","笑cry","漫威",
             "清新","匠心","VR","心理学","舞蹈","品牌广告","粉丝自制","电影相关","萝莉","魔性"
             ,"第一视角","教程","毕业设计","奥斯卡","燃","冰与火之歌","温情","线下campaign","公益")
-    lateinit var mRootView: View
     lateinit var mCircularRevealAnim: CircularRevealAnim
     lateinit var mAdatper : SearchAdapter
+
+
+    private var _binding: SearchFragmentBinding? = null
+    private val binding get() = _binding!!
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setStyle(DialogFragment.STYLE_NO_FRAME, R.style.DialogStyle);
@@ -53,10 +57,9 @@ class SearchFragment : DialogFragment(), CircularRevealAnim.AnimListener,
         window.setWindowAnimations(R.style.DialogEmptyAnimation)//取消过渡动画 , 使DialogSearch的出现更加平滑
     }
 
-    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        mRootView = inflater?.inflate(R.layout.search_fragment, container, false)!!
-
-        return mRootView
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        _binding =SearchFragmentBinding.inflate(inflater)
+        return binding.root
     }
 
 
@@ -65,7 +68,12 @@ class SearchFragment : DialogFragment(), CircularRevealAnim.AnimListener,
         init()
         setData()
     }
-
+    
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+    
     private fun setData() {
         mAdatper = SearchAdapter(activity, data as ArrayList<String>)
         mAdatper.setOnDialogDismissListener(object :SearchAdapter.onDialogDismiss{
@@ -73,40 +81,40 @@ class SearchFragment : DialogFragment(), CircularRevealAnim.AnimListener,
                 hideAnim()
             }
         })
-        val manager = FlexboxLayoutManager()
+        val manager = FlexboxLayoutManager(activity)
         //设置主轴排列方式
         manager.flexDirection = FlexDirection.ROW
         //设置是否换行
         manager.flexWrap = FlexWrap.WRAP
-        recyclerView.layoutManager = manager
-        recyclerView.itemAnimator = DefaultItemAnimator()
-        recyclerView.adapter = mAdatper
+        binding.recyclerView.layoutManager = manager
+        binding.recyclerView.itemAnimator = androidx.recyclerview.widget.DefaultItemAnimator()
+        binding.recyclerView.adapter = mAdatper
     }
 
     private fun init() {
-        tv_hint.typeface = Typeface.createFromAsset(activity.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
+        binding.tvHint.typeface = Typeface.createFromAsset(activity.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
         mCircularRevealAnim = CircularRevealAnim()
         mCircularRevealAnim.setAnimListener(this)
         dialog.setOnKeyListener(this)
-        iv_search_search.viewTreeObserver.addOnPreDrawListener(this)
-        iv_search_search.setOnClickListener(this)
-        iv_search_back.setOnClickListener(this)
+        binding.ivSearchSearch.viewTreeObserver.addOnPreDrawListener(this)
+        binding.ivSearchSearch.setOnClickListener(this)
+        binding.ivSearchBack.setOnClickListener(this)
     }
 
     override fun onHideAnimationEnd() {
-        et_search_keyword.setText("");
+        binding.etSearchKeyword.setText("");
         dismiss();
     }
 
     override fun onShowAnimationEnd() {
         if (isVisible) {
-            KeyBoardUtils.openKeyboard(activity, et_search_keyword);
+            KeyBoardUtils.openKeyboard(activity, binding.etSearchKeyword);
         }
     }
 
     override fun onPreDraw(): Boolean {
-        iv_search_search.viewTreeObserver.removeOnPreDrawListener(this);
-        mCircularRevealAnim.show(iv_search_search, mRootView);
+        binding.ivSearchSearch.viewTreeObserver.removeOnPreDrawListener(this);
+        mCircularRevealAnim.show(binding.ivSearchSearch, binding.root);
         return true;
     }
 
@@ -120,12 +128,12 @@ class SearchFragment : DialogFragment(), CircularRevealAnim.AnimListener,
     }
 
     private fun search() {
-        val searchKey = et_search_keyword.text.toString()
+        val searchKey = binding.etSearchKeyword.text.toString()
         if (TextUtils.isEmpty(searchKey.trim({ it <= ' ' }))) {
             Toast.makeText(activity, "请输入关键字", Toast.LENGTH_SHORT).show()
         } else {
             hideAnim()
-            var keyWord = et_search_keyword.text.toString().trim()
+            var keyWord = binding.etSearchKeyword.text.toString().trim()
             var intent : Intent = Intent(activity, ResultActivity::class.java)
             intent.putExtra("keyWord",keyWord)
             activity?.startActivity(intent)
@@ -133,8 +141,8 @@ class SearchFragment : DialogFragment(), CircularRevealAnim.AnimListener,
     }
 
     private fun hideAnim() {
-        KeyBoardUtils.closeKeyboard(activity, et_search_keyword);
-        mCircularRevealAnim.hide(iv_search_search, mRootView)
+        KeyBoardUtils.closeKeyboard(activity, binding.etSearchKeyword);
+        mCircularRevealAnim.hide(binding.ivSearchSearch, binding.root)
     }
 
     override fun onClick(v: View?) {
