@@ -4,23 +4,25 @@ import android.os.AsyncTask
 import android.os.Bundle
 import android.os.Handler
 import android.os.Message
-import android.support.v7.app.AlertDialog
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
 import android.view.View
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
+import com.dylanc.viewbinding.inflate
 import com.gyf.barlibrary.ImmersionBar
 import com.tt.lvruheng.eyepetizer.R
 import com.tt.lvruheng.eyepetizer.adapter.DownloadAdapter
+import com.tt.lvruheng.eyepetizer.databinding.ActivityWatchBinding
 import com.tt.lvruheng.eyepetizer.mvp.model.bean.VideoBean
 import com.tt.lvruheng.eyepetizer.utils.ObjectSaveUtils
 import com.tt.lvruheng.eyepetizer.utils.SPUtils
-import kotlinx.android.synthetic.main.activity_watch.*
 import zlc.season.rxdownload2.RxDownload
 
 /**
  * Created by lvruheng on 2017/7/12.
  */
 class CacheActivity : AppCompatActivity() {
+    private val binding: ActivityWatchBinding by inflate()
+
     var mList = ArrayList<VideoBean>()
     lateinit var mAdapter: DownloadAdapter
     var mHandler: Handler = object : Handler() {
@@ -28,9 +30,9 @@ class CacheActivity : AppCompatActivity() {
             super.handleMessage(msg)
             var list = msg?.data?.getParcelableArrayList<VideoBean>("beans")
             if (list?.size?.compareTo(0) == 0) {
-                tv_hint.visibility = View.VISIBLE
+                binding.tvHint.visibility = View.VISIBLE
             } else {
-                tv_hint.visibility = View.GONE
+                binding.tvHint.visibility = View.GONE
                 if (mList.size > 0) {
                     mList.clear()
                 }
@@ -47,7 +49,7 @@ class CacheActivity : AppCompatActivity() {
         ImmersionBar.with(this).transparentBar().barAlpha(0.3f).fitsSystemWindows(true).init()
         setToolbar()
         DataAsyncTask(mHandler, this).execute()
-        recyclerView.layoutManager = LinearLayoutManager(this)
+        binding.recyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(this)
         mAdapter = DownloadAdapter(this, mList)
         mAdapter.setOnLongClickListener(object : DownloadAdapter.OnLongClickListener {
             override fun onLongClick(position: Int) {
@@ -55,7 +57,7 @@ class CacheActivity : AppCompatActivity() {
             }
         })
 
-        recyclerView.adapter = mAdapter
+        binding.recyclerView.adapter = mAdapter
     }
 
     private fun addDialog(position: Int) {
@@ -83,11 +85,11 @@ class CacheActivity : AppCompatActivity() {
     }
 
     private fun setToolbar() {
-        setSupportActionBar(toolbar)
+        setSupportActionBar(binding.toolbar)
         var bar = supportActionBar
         bar?.title = "我的缓存"
         bar?.setDisplayHomeAsUpEnabled(true)
-        toolbar.setNavigationOnClickListener {
+        binding.toolbar.setNavigationOnClickListener {
             onBackPressed()
         }
     }
