@@ -1,16 +1,15 @@
 package com.tt.lvruheng.eyepetizer.ui
 
 import android.graphics.Typeface
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.util.Log
 import android.view.KeyEvent
 import android.view.View
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import com.dylanc.viewbinding.inflate
 import com.gyf.barlibrary.ImmersionBar
 import com.tt.lvruheng.eyepetizer.R
+import com.tt.lvruheng.eyepetizer.databinding.ActivityMainBinding
 import com.tt.lvruheng.eyepetizer.search.SEARCH_TAG
 import com.tt.lvruheng.eyepetizer.search.SearchFragment
 import com.tt.lvruheng.eyepetizer.ui.fragment.FindFragment
@@ -18,17 +17,16 @@ import com.tt.lvruheng.eyepetizer.ui.fragment.HomeFragment
 import com.tt.lvruheng.eyepetizer.ui.fragment.HotFragment
 import com.tt.lvruheng.eyepetizer.ui.fragment.MineFragment
 import com.tt.lvruheng.eyepetizer.utils.showToast
-import kotlinx.android.synthetic.main.activity_main.*
 import java.util.*
-import android.view.WindowManager
-
 
 
 class MainActivity : AppCompatActivity(), View.OnClickListener {
-    var homeFragment: HomeFragment? = null
-    var findFragment: FindFragment? = null
-    var hotFragemnt: HotFragment? = null
-    var mineFragment: MineFragment? = null
+    private val binding: ActivityMainBinding by inflate()
+
+    lateinit var homeFragment: HomeFragment
+    lateinit var findFragment: FindFragment
+    lateinit var hotFragemnt: HotFragment
+    lateinit var mineFragment: MineFragment
     var mExitTime: Long = 0
     var toast: Toast? = null
     lateinit var searchFragment: SearchFragment
@@ -47,10 +45,10 @@ class MainActivity : AppCompatActivity(), View.OnClickListener {
 
     private fun initToolbar() {
         var today = getToday()
-        tv_bar_title.text = today
-        tv_bar_title.typeface = Typeface.createFromAsset(this.assets, "fonts/Lobster-1.4.otf")
-        iv_search.setOnClickListener {
-            if (rb_mine.isChecked) {
+        binding.tvBarTitle.text = today
+        binding.tvBarTitle.typeface = Typeface.createFromAsset(this.assets, "fonts/Lobster-1.4.otf")
+        binding.ivSearch.setOnClickListener {
+            if (binding.rbMine.isChecked) {
                 //todo 点击设置
             } else {
                 //todo 点击搜索
@@ -76,7 +74,7 @@ class MainActivity : AppCompatActivity(), View.OnClickListener {
     private fun initFragment(savedInstanceState: Bundle?) {
         if (savedInstanceState != null) {
             //异常情况
-            val mFragments: List<Fragment> = supportFragmentManager.fragments
+            val mFragments: List<androidx.fragment.app.Fragment> = supportFragmentManager.fragments
             for (item in mFragments) {
                 if (item is HomeFragment) {
                     homeFragment = item
@@ -111,74 +109,74 @@ class MainActivity : AppCompatActivity(), View.OnClickListener {
     }
 
     private fun setRadioButton() {
-        rb_home.isChecked = true
-        rb_home.setTextColor(resources.getColor(R.color.black))
-        rb_home.setOnClickListener(this)
-        rb_find.setOnClickListener(this)
-        rb_hot.setOnClickListener(this)
-        rb_mine.setOnClickListener(this)
+        binding.rbHome.isChecked = true
+        binding.rbHome.setTextColor(resources.getColor(R.color.black))
+        binding.rbHome.setOnClickListener(this)
+        binding.rbFind.setOnClickListener(this)
+        binding.rbHot.setOnClickListener(this)
+        binding.rbMine.setOnClickListener(this)
     }
 
     override fun onClick(v: View?) {
         clearState()
         when (v?.id) {
             R.id.rb_find -> {
-                rb_find.isChecked = true
-                rb_find.setTextColor(resources.getColor(R.color.black))
+                binding.rbFind.isChecked = true
+                binding.rbFind.setTextColor(resources.getColor(R.color.black))
                 supportFragmentManager.beginTransaction().show(findFragment)
                         .hide(homeFragment)
                         .hide(mineFragment)
                         .hide(hotFragemnt)
                         .commit()
-                tv_bar_title.text = "Discover"
-                tv_bar_title.visibility = View.VISIBLE
-                iv_search.setImageResource(R.drawable.icon_search)
+                binding.tvBarTitle.text = "Discover"
+                binding.tvBarTitle.visibility = View.VISIBLE
+                binding.ivSearch.setImageResource(R.drawable.icon_search)
             }
             R.id.rb_home -> {
-                rb_home.isChecked = true
-                rb_home.setTextColor(resources.getColor(R.color.black))
+                binding.rbHome.isChecked = true
+                binding.rbHome.setTextColor(resources.getColor(R.color.black))
                 supportFragmentManager.beginTransaction().show(homeFragment)
                         .hide(findFragment)
                         .hide(mineFragment)
                         .hide(hotFragemnt)
                         .commit()
-                tv_bar_title.text = getToday()
-                tv_bar_title.visibility = View.VISIBLE
-                iv_search.setImageResource(R.drawable.icon_search)
+                binding.tvBarTitle.text = getToday()
+                binding.tvBarTitle.visibility = View.VISIBLE
+                binding.ivSearch.setImageResource(R.drawable.icon_search)
             }
             R.id.rb_hot -> {
-                rb_hot.isChecked = true
-                rb_hot.setTextColor(resources.getColor(R.color.black))
+                binding.rbHot.isChecked = true
+                binding.rbHot.setTextColor(resources.getColor(R.color.black))
                 supportFragmentManager.beginTransaction().show(hotFragemnt)
                         .hide(findFragment)
                         .hide(mineFragment)
                         .hide(homeFragment)
                         .commit()
-                tv_bar_title.text = "Ranking"
-                tv_bar_title.visibility = View.VISIBLE
-                iv_search.setImageResource(R.drawable.icon_search)
+                binding.tvBarTitle.text = "Ranking"
+                binding.tvBarTitle.visibility = View.VISIBLE
+                binding.ivSearch.setImageResource(R.drawable.icon_search)
             }
             R.id.rb_mine -> {
-                rb_mine.isChecked = true
-                rb_mine.setTextColor(resources.getColor(R.color.black))
+                binding.rbMine.isChecked = true
+                binding.rbMine.setTextColor(resources.getColor(R.color.black))
                 supportFragmentManager.beginTransaction().show(mineFragment)
                         .hide(findFragment)
                         .hide(homeFragment)
                         .hide(hotFragemnt)
                         .commit()
-                tv_bar_title.visibility = View.INVISIBLE
-                iv_search.setImageResource(R.drawable.icon_setting)
+                binding.tvBarTitle.visibility = View.INVISIBLE
+                binding.ivSearch.setImageResource(R.drawable.icon_setting)
             }
         }
 
     }
 
     private fun clearState() {
-        rg_root.clearCheck()
-        rb_home.setTextColor(resources.getColor(R.color.gray))
-        rb_mine.setTextColor(resources.getColor(R.color.gray))
-        rb_hot.setTextColor(resources.getColor(R.color.gray))
-        rb_find.setTextColor(resources.getColor(R.color.gray))
+        binding.rgRoot.clearCheck()
+        binding.rbHome.setTextColor(resources.getColor(R.color.gray))
+        binding.rbMine.setTextColor(resources.getColor(R.color.gray))
+        binding.rbHot.setTextColor(resources.getColor(R.color.gray))
+        binding.rbFind.setTextColor(resources.getColor(R.color.gray))
     }
 
     override fun onPause() {
