@@ -3,28 +3,28 @@ package com.tt.lvruheng.eyepetizer.ui
 import android.content.Context
 import android.content.pm.ActivityInfo
 import android.content.res.Configuration
-import android.graphics.Typeface
-import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import com.tt.lvruheng.eyepetizer.R
-import com.tt.lvruheng.eyepetizer.mvp.model.bean.VideoBean
-import kotlinx.android.synthetic.main.activity_video_detail.*
 import android.graphics.BitmapFactory
-import com.bumptech.glide.Glide
+import android.graphics.Typeface
 import android.os.AsyncTask
+import android.os.Bundle
 import android.os.Handler
 import android.os.Message
 import android.util.Log
 import android.view.View
 import android.widget.ImageView
+import androidx.appcompat.app.AppCompatActivity
+import com.bumptech.glide.Glide
+import com.dylanc.viewbinding.inflate
 import com.shuyu.gsyvideoplayer.GSYVideoPlayer
 import com.shuyu.gsyvideoplayer.utils.OrientationUtils
 import com.shuyu.gsyvideoplayer.video.StandardGSYVideoPlayer
+import com.tt.lvruheng.eyepetizer.R
+import com.tt.lvruheng.eyepetizer.databinding.ActivityVideoDetailBinding
+import com.tt.lvruheng.eyepetizer.mvp.model.bean.VideoBean
 import com.tt.lvruheng.eyepetizer.utils.*
 import zlc.season.rxdownload2.RxDownload
 import java.io.FileInputStream
 import java.io.FileNotFoundException
-import java.net.URI
 import java.util.concurrent.ExecutionException
 
 
@@ -32,6 +32,8 @@ import java.util.concurrent.ExecutionException
  * Created by lvruheng on 2017/7/7.
  */
 class VideoDetailActivity : AppCompatActivity() {
+    private val binding: ActivityVideoDetailBinding by inflate()
+
     companion object {
         var MSG_IMAGE_LOADED = 101
     }
@@ -48,7 +50,7 @@ class VideoDetailActivity : AppCompatActivity() {
             when (msg?.what) {
                 MSG_IMAGE_LOADED -> {
                     Log.e("video", "setImage")
-                    gsy_player.setThumbImageView(imageView)
+                    binding.gsyPlayer.setThumbImageView(imageView)
                 }
             }
         }
@@ -64,11 +66,11 @@ class VideoDetailActivity : AppCompatActivity() {
 
     private fun initView() {
         var bgUrl = bean.blurred
-        bgUrl?.let { ImageLoadUtils.displayHigh(this, iv_bottom_bg, bgUrl) }
-        tv_video_desc.text = bean.description
-        tv_video_desc.typeface = Typeface.createFromAsset(this.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
-        tv_video_title.text = bean.title
-        tv_video_title.typeface = Typeface.createFromAsset(this.assets, "fonts/FZLanTingHeiS-L-GB-Regular.TTF")
+        bgUrl?.let { ImageLoadUtils.displayHigh(this, binding.ivBottomBg, bgUrl) }
+        binding.tvVideoDesc.text = bean.description
+        binding.tvVideoDesc.typeface = Typeface.createFromAsset(this.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
+        binding.tvVideoTitle.text = bean.title
+        binding.tvVideoTitle.typeface = Typeface.createFromAsset(this.assets, "fonts/FZLanTingHeiS-L-GB-Regular.TTF")
         var category = bean.category
         var duration = bean.duration
         var minute = duration?.div(60)
@@ -85,11 +87,11 @@ class VideoDetailActivity : AppCompatActivity() {
         } else {
             realSecond = second.toString()
         }
-        tv_video_time.text = "$category / $realMinute'$realSecond''"
-        tv_video_favor.text = bean.collect.toString()
-        tv_video_share.text = bean.share.toString()
-        tv_video_reply.text = bean.share.toString()
-        tv_video_download.setOnClickListener {
+        binding.tvVideoTitle.text = "$category / $realMinute'$realSecond''"
+        binding.tvVideoFavor.text = bean.collect.toString()
+        binding.tvVideoShare.text = bean.share.toString()
+        binding.tvVideoReply.text = bean.share.toString()
+        binding.tvVideoDownload.setOnClickListener {
             //点击下载
             var url = bean.playUrl?.let { it1 -> SPUtils.getInstance(this, "downloads").getString(it1) }
             if (url.equals("")) {
@@ -122,30 +124,30 @@ class VideoDetailActivity : AppCompatActivity() {
         var uri = intent.getStringExtra("loaclFile")
         if(uri!=null){
             Log.e("uri",uri)
-            gsy_player.setUp(uri, false, null, null)
+            binding.gsyPlayer.setUp(uri, false, null, null)
         }else{
-            gsy_player.setUp(bean.playUrl, false, null, null)
+            binding.gsyPlayer.setUp(bean.playUrl, false, null, null)
         }
         //增加封面
         imageView = ImageView(this)
         imageView.scaleType = ImageView.ScaleType.CENTER_CROP
         ImageViewAsyncTask(mHandler, this, imageView).execute(bean.feed)
-        gsy_player.titleTextView.visibility = View.GONE
-        gsy_player.backButton.visibility = View.VISIBLE
-        orientationUtils = OrientationUtils(this, gsy_player)
-        gsy_player.setIsTouchWiget(true);
+        binding.gsyPlayer.titleTextView.visibility = View.GONE
+        binding.gsyPlayer.backButton.visibility = View.VISIBLE
+        orientationUtils = OrientationUtils(this, binding.gsyPlayer)
+        binding.gsyPlayer.setIsTouchWiget(true);
         //关闭自动旋转
-        gsy_player.isRotateViewAuto = false;
-        gsy_player.isLockLand = false;
-        gsy_player.isShowFullAnimation = false;
-        gsy_player.isNeedLockFull = true;
-        gsy_player.fullscreenButton.setOnClickListener {
+        binding.gsyPlayer.isRotateViewAuto = false;
+        binding.gsyPlayer.isLockLand = false;
+        binding.gsyPlayer.isShowFullAnimation = false;
+        binding.gsyPlayer.isNeedLockFull = true;
+        binding.gsyPlayer.fullscreenButton.setOnClickListener {
             //直接横屏
             orientationUtils.resolveByClick();
             //第一个true是否需要隐藏actionbar，第二个true是否需要隐藏statusbar
-            gsy_player.startWindowFullscreen(mContext, true, true);
+            binding.gsyPlayer.startWindowFullscreen(mContext, true, true);
         }
-        gsy_player.setStandardVideoAllCallBack(object : VideoListener() {
+        binding.gsyPlayer.setStandardVideoAllCallBack(object : VideoListener() {
             override fun onPrepared(url: String?, vararg objects: Any?) {
                 super.onPrepared(url, *objects)
                 //开始播放了才能旋转和全屏
@@ -167,11 +169,11 @@ class VideoDetailActivity : AppCompatActivity() {
                 orientationUtils?.let { orientationUtils.backToProtVideo() }
             }
         })
-        gsy_player.setLockClickListener { view, lock ->
+        binding.gsyPlayer.setLockClickListener { view, lock ->
             //配合下方的onConfigurationChanged
             orientationUtils.isEnable = !lock
         }
-        gsy_player.backButton.setOnClickListener(View.OnClickListener {
+        binding.gsyPlayer.backButton.setOnClickListener(View.OnClickListener {
             onBackPressed()
         })
 
@@ -241,16 +243,16 @@ class VideoDetailActivity : AppCompatActivity() {
         }
     }
 
-    override fun onConfigurationChanged(newConfig: Configuration?) {
+    override fun onConfigurationChanged(newConfig: Configuration) {
         super.onConfigurationChanged(newConfig)
         if (isPlay && !isPause) {
             if (newConfig?.orientation == ActivityInfo.SCREEN_ORIENTATION_USER) {
-                if (!gsy_player.isIfCurrentIsFullscreen) {
-                    gsy_player.startWindowFullscreen(mContext, true, true)
+                if (!binding.gsyPlayer.isIfCurrentIsFullscreen) {
+                    binding.gsyPlayer.startWindowFullscreen(mContext, true, true)
                 }
             } else {
                 //新版本isIfCurrentIsFullscreen的标志位内部提前设置了，所以不会和手动点击冲突
-                if (gsy_player.isIfCurrentIsFullscreen) {
+                if (binding.gsyPlayer.isIfCurrentIsFullscreen) {
                     StandardGSYVideoPlayer.backFromWindowFull(this);
                 }
                 orientationUtils?.let { orientationUtils.isEnable = true }
