@@ -6,19 +6,23 @@ import android.view.View
 import android.widget.AdapterView
 import com.tt.lvruheng.eyepetizer.R
 import com.tt.lvruheng.eyepetizer.adapter.FindAdapter
+import com.tt.lvruheng.eyepetizer.databinding.FindFragmentBinding
+import com.tt.lvruheng.eyepetizer.databinding.SearchFragmentBinding
 import com.tt.lvruheng.eyepetizer.mvp.contract.FindContract
 import com.tt.lvruheng.eyepetizer.mvp.model.bean.FindBean
 import com.tt.lvruheng.eyepetizer.mvp.presenter.FindPresenter
 import com.tt.lvruheng.eyepetizer.ui.FindDetailActivity
-import kotlinx.android.synthetic.main.find_fragment.*
+
 
 /**
  * Created by lvruheng on 2017/7/4.
  */
-class FindFragment : BaseFragment(),FindContract.View {
+class FindFragment : BaseFragment<FindFragmentBinding>(),FindContract.View {
     var mPresenter: FindPresenter? = null
     var mAdapter : FindAdapter? = null
     var mList : MutableList<FindBean>? = null
+
+
     override fun setData(beans: MutableList<FindBean>) {
         mAdapter?.mList = beans
         mList = beans
@@ -33,12 +37,13 @@ class FindFragment : BaseFragment(),FindContract.View {
         mPresenter = FindPresenter(context,this)
         mPresenter?.start()
         mAdapter = FindAdapter(context,mList)
-        gv_find.adapter = mAdapter
-        gv_find.setOnItemClickListener { parent, view, position, id ->
-            var bean = mList?.get(position)
-            var name = bean?.name
-            var intent : Intent = Intent(context,FindDetailActivity::class.java)
+        binding.gvFind.adapter = mAdapter
+        binding.gvFind.setOnItemClickListener { parent, view, position, id ->
+            val bean = mList?.get(position)
+            val name = bean?.name
+            val intent = Intent(context,FindDetailActivity::class.java)
             intent.putExtra("name",name)
+            intent.putExtra("id",bean?.id)
             startActivity(intent)
 
         }
