@@ -4,15 +4,15 @@ import android.content.Intent
 import android.graphics.Typeface
 import android.view.View
 import com.tt.lvruheng.eyepetizer.R
+import com.tt.lvruheng.eyepetizer.databinding.MineFragmentBinding
 import com.tt.lvruheng.eyepetizer.ui.AdviseActivity
 import com.tt.lvruheng.eyepetizer.ui.CacheActivity
 import com.tt.lvruheng.eyepetizer.ui.WatchActivity
-import kotlinx.android.synthetic.main.mine_fragment.*
 
 /**
  * Created by lvruheng on 2017/7/4.
  */
-class MineFragment : BaseFragment(),View.OnClickListener{
+class MineFragment : BaseFragment<MineFragmentBinding>(),View.OnClickListener{
     override fun onClick(v: View?) {
         when(v?.id){
             R.id.tv_watch ->{
@@ -35,12 +35,12 @@ class MineFragment : BaseFragment(),View.OnClickListener{
     }
 
     override fun initView() {
-        tv_advise.setOnClickListener(this)
-        tv_watch.setOnClickListener(this)
-        tv_save.setOnClickListener(this)
-        tv_advise.typeface = Typeface.createFromAsset(context?.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
-        tv_watch.typeface = Typeface.createFromAsset(context?.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
-        tv_save.typeface = Typeface.createFromAsset(context?.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
+        binding.tvAdvise.setOnClickListener(this)
+        binding.tvWatch.setOnClickListener(this)
+        binding.tvSave.setOnClickListener(this)
+        binding.tvAdvise.typeface = Typeface.createFromAsset(context?.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
+        binding.tvWatch.typeface = Typeface.createFromAsset(context?.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
+        binding.tvSave.typeface = Typeface.createFromAsset(context?.assets, "fonts/FZLanTingHeiS-DB1-GB-Regular.TTF")
     }
 
 }
